requirements.txt
gradio>=4.0.0
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
scikit-learn>=1.3.0
plotly>=5.18.0 